move separate function into module pastis_functions.py
